Rename compilerFlag parameter to compilationCondition
